Requirements and bugfix flask

+++ Flask/Flask.py
@@ -34,14 +34,14 @@ def cleanup_files():
     csv_files_path = config.PATH_TO_CSV_FILES
     for csv_file in os.listdir(csv_files_path):
         csv_file_path = os.path.join(csv_files_path, csv_file)
-        if os.path.isfile(csv_file_path):
+        if os.path.isfile(csv_file_path) and item != '.gitkeep':
             os.remove(csv_file_path)
 
     # Clear contents of PATH_TO_RESULTS
     results_path = config.PATH_TO_RESULTS
     for result_item in os.listdir(results_path):
         result_item_path = os.path.join(results_path, result_item)
-        if os.path.isfile(result_item_path):
+        if os.path.isfile(result_item_path) and item != '.gitkeep':
             os.remove(result_item_path)
         elif os.path.isdir(result_item_path):
             shutil.rmtree(result_item_path)
